XSPH Correction + Others
Added the XSPH correction, plus a density reinitialisation (doesn't seem to change anything in particular), that have improved the physics a bit from first appearances. Care needs to be taken with the XSPH. the recommended parameter of 0.5 led to particles converging to a single point and a SEGV from the neighbour search. Currently set at 0.1, and that's stable. Also added error calculation at each timestep, which is highly useful to see if the solution is diverging. Possibly calculate the energy of the system? Look at conservation of that?